fix(triggers): fix handling of blank regex in docker registry triggers
When you remove a previously existing regex from the trigger in the
UI, deck keeps a tag field with an empty string for the value in the
json. This arrives as non null to the DockerEventMonitor, which  causes
the trigger to not fire for any new tag until you manually update the
pipeline json to remove the field. This change makes the logic more
resilient to malformed input. Blank or empty tag regex doesn't make
sense so it should behave the same as when the regex is null.