refactor(core): Push shared functionality to TriggerMonitor.java
There is a lot of duplicated code in the subclasses of TriggerMonitor;
specifically a number of member variables are defined on all subclasses
and can be pushed to the base class. Similarly, the core logic in
processEvent is the same for all subclasses and should live in the
base class.