refactor(core): Split pipeline matching and triggering logic
The EventMonitor classes are currently responsible for looking
up the pipelines from the cache, figuring out which ones should be
triggered by a given event, and actually initiating the resulting
pipelines. This leads so some reasonably complex inheritance
relationships that will only become more complex as we add more
artifact extraction logic to echo.

Factor out the logic to match events to pipelines into a separate
class hierarchy that is composed into the event monitor. The event
monitor is then responsible for looking up the pipelines and starting
the matching ones, but delegates responsiblity for actually determining
which pipelines match.