refactor(core): Use generics in echo event handlers
There is a lot of casting in the echo event handlers, as they
all implement a common base class that only specifies the
event in question is a TriggerEvent. This change makes each
event handler parameterized by the type of event it's handling
so we can get type safety and avoid casting.