feat(core): Add functionality for artifact-extracting Jinja templates
Add functionality for adding Jinja templates that will be used to
extract artifacts from pipeline triggers. In particular, adds an enum
of standard templates included in the jar as well as the ability to
configure custom templates via echo config.

I've tried to make this general enough that we can in the future
refactor some of the duplicated code in the google/aws pubsub
template handling to use this, but have not actually taken on that
refactor in this commit.