feat(artifacts): Use property to select template
As a first pass, let's allow users to select a Jinja template
to extract artifacts from a Jenkins job by exporting the
template to use as a property. We'll likely want to provide
other ways of configuring this in the future, but this
is a lightweight way to get started, and will allow users
to start using this feature.