fix(core): Don't retry pipeline initiation when receiving a 400
When a request to Orca to start a pipeline fails, Echo currently
retries a number of times (defaulting to 5). In cases where the
failure is due to a problem with the request, we should not retry;
gate the retry logic so that we don't retry when receiving a
400 - Bad Request from Orca. In particular, this will stop having
pipelines missing artifacts lead to 5 failed executions in the history.