fix(echo): relax healthcheck constraints
To avoid cases where an otherwise good echo instance knocks itself out
of service and stops processing events when its last front50 poll is a
bit stale, we only ever allow its health to transition from DOWN to UP
as soon as a successful cycle has completed.

The previous setting was quite aggressive, going DOWN as early as after
2 missed front50 polling cycles. Even if slightly degraded, it should
still be ok in most cases to keep processing events based on data that's
a few minutes old, and alert otherwise.