fix(artifacts): Move build fetching to build event handler
Artifact contstraints on build triggers currently don't work as
the constraint is evaluated while determining which trigger to
use but build information and artifact inflation only happens
once the pipeline is ready to be sent to Orca. Move the fetching
of build information to earlier in the triggering process (and
scope it to only build triggers). A follow-up change will move
artifact inflation earlier as well.