refactor(artifacts): Remove jinja logic and import from kork
This creates some base classes for extracting artifacts from
messages using Jinja that will be pushed into kork, and updates
echo to use these. It also better encapsulates the dependencies
of the artifact extractors so that we only need to pass around
a factory bean rather than all of the internal dependencies.