Boot2 (#516)
* chore(dependencies): Autobump spinnaker-dependencies (#506)

* fix(MPT): Fixes triggering for template triggers. (#504)

Echo has its own cache of pipeline configs, but
does not hydrate templated pipeline configs into
the full runnable pipeline. Hence if you define
a trigger in the template, Echo will not trigger it
when it receives a matching trigger event.

* feat(buildmasters): Permission support for build masters (CI's) (#467)

* feat(buildmasters): Permission support for build masters (CI's)
* Propagate auth to Igor

* chore(dependencies): Autobump spinnaker-dependencies (#507)

* feat(artifacts): populate artifacts for manual trigger (#509)

* fix(pipeline): if map -> Pipeline fails, log and move on (#510)

* fix(events): Guard against NPE when artifacts is not set. (#511)

* fix(mpt): fix planing of v1 pipelines for manual triggers (#512)

* fix(triggers): Better error handling for trigger failures (#513)

If planning a pipeline fails, make sure to log a metric

* refactor(logging): Log better messages when events fail (#514)

Currently we log the whole event first which can be quite large and thus overloads our log provider (truncating the message)
Split the log into parts such as the important bits are first and the rest can get truncated

* fix(retrofit): make sure all retrofit clients use Jackson (#515)

GSON is the default converter. It, for some reason, takes a perfectly valid JSON
with integers and parses them as doubles (sometimes). Jackson doesn't do that.
Jackson is better, use it.

* chore(deps): adding boot2 deps

* chore(boot2): update all annotation values

* fix(test): change to jupiter

* fix(boot2): disable gson auto config

* chore(deps): upgrade spin-dep to have upgraded kork