fix(triggers): fetch latest config from front50 on impending trigger
Fixes the following issue:
* save a pipeline through deck, wait for the UI to display "In Sync With
Server" (as reported by orca's MonitorFront50Task)
* trigger the pipeline manually after a short period of time
* the execution actually references the old pipeline configuration

This is because echo references the pipeline config from its in-memory
cache, not the version that was most recently saved to front50.

The window of inconsistency could be arbitrarily long, depending on echo's
PipelineCache polling configuration and front50's own caching.

This change fixes the issue by fetching the latest config from front50
just before the pipeline execution request is submitted to orca (in
TriggerMonitor::triggerMatchingPipelines)