feat(auth): Propagate `X-SPINNAKER-ACCOUNTS` when triggering pipelines
Echo and externally triggered (via `gate`) pipelines differ in that the
former does not propagate `X-SPINNAKER-ACCOUNTS`, even when a service
account is being used.

When `services.fiat.legacyFallback` is enabled, this PR will fetch
the allowed accounts for either the specified service account name
or `anonymous`.

The goal is to make fallbacks consistent regardless of how a pipeline
was triggered.