feat(gcb): Call igor to extract artifacts when triggering
Extracting artifacts using a Jinja template does not properly
work when a GCB build produces GCS artifacts. This is because
the build saves a manifest file to GCS, and we need to read the
manifest file to find the location of each produced artifact.

The functionality to do this was recently added to igor; if
we're listening to a GCB pubsub topic, use a custom artifact
extractor that asks igor to get the artifacts from the build.