fix(logging): add fabulous logging for when pipelines fail to trigger
The impetus for this change is to be able to better alert on and understand pipeline trigger failures.
We've been bit by trigger failures that went undiscovered (sometimes for days) until our users complained.
With this change we can reliable alert on `orca.trigger.errors` metric which is incremented when kicking
of a pipeline fails (unlike the previous metric `orca.error` which was sometimes triggered erroneously).

When triggering does fail, this change will log, in much more detail, why the trigger failed as well as
log the full payload submitted to `orca` for quick and easy debugging/reproducability of the issue.

Bonus 1: This change (mostly) removes the reliance on Rx and replaces it with an executor to
trigger pipelines.

Bonus 2: Fixup calls to `front50` to allow them to be anonymous