fix(bitbucket): Log when Bitbucket event unknown
There are rare cases (often when the /bitbucket endpoint is used instead
of /stash) where the bitbucket event type doesn't match _either_ the
Cloud or Server types; this can cause an NPE in the subsequent log
message because "branch" doesn't get set (this was also fixed, although
should probably not happen any more).  In this case, we log the fact
that we couldn't determine the Cloud/Server type based on the event_type
(noting the event_type we _did_ get) and return.