fix(triggers): Fix NPE on null application/name
If the name or application on a trigger is null, we throw
an NPE.  This only happens in the 1.12.x branch, as in 1.13
and beyond the @NonNull annotation on these fields is causing
invalid pipelines to fail to deserialize and we thus never
check them against triggers.

Add null checks to the 1.12 branch directly to fix the issue here.
I'd rather not add these checks in later branches as having the
field annotated as @NonNull is a better fix and it's redundant to
perform a lot of null checks against @NonNull fields.