fix(notifications):  Need to maintain enum order
HIPCHAT was removed from the Enum here, but this enum is duplicated in
other microservices (for example, in Orca,
orca-echo/src/main/groovy/com/netflix/spinnaker/orca/echo/EchoService.groovy
has a copy of the list)

This is a quick fix to repair the numbering scheme.  This should
probably be fully refactored so the enum is either a) specifying actual
values for each type so they maintain consistency going forward, b)
unify in kork somehow, or c) both.