fix(trigger): handle quartz errors better
We want to know when `echo` fails to register a trigger with quartz but only if it's not a user error (e.g. bad cron express, cron expression in the past, etc).
Quartz, unfortunately, often throws a non-descript `SchedulerException` which doesn't provide machine identifiable way to decipher if this is a user error or a system error.

This change replicates some logic that's in [quartz](https://github.com/quartz-scheduler/quartz/blob/939a9a25fbf3b374608de59e413405aebe10c089/quartz-core/src/main/java/org/quartz/core/QuartzScheduler.java#L882-L896)
to decide if the trigger is valid, if not we will log a warning, but won't increment any failure metrics.
If the are exceptions/errors outside of the trigger validation path, those will still log errors and increment failure metrics