fix(webhooks/stash): Adding jsonalias for refChanges field of StashWebhookEvent to account for Stash webhook payloads that follow this format.
This patch adds a simple json alias for the refChanges field of StashWebhookEvent to account for payloads that have a slightly different format of changes for the field name. Several issues have been reported for Stash triggers not functioning but have been unable to tie a reported issue to this field in particular. Currently, I was struggling with Stash triggers and built a custom image with this proposed fix to validate. This fix corrected Stash triggers in my case and has no impact to existing payloads that follow the refChanges format.