feat(pipelinetriggers): only cache enabled pipelines with enabled triggers of specific types
when pipelineCache.filterFront50Pipelines is true.  Determine the specific types from the
types that the available BaseTriggerEventHandler beans support, with cron triggers as
well. Adjust the logic in ManualEventHandler to work when the cache only contains a subset
of pipelines.

When pipelineCache.filterFront50Pipelines is false, retain the current behavior to query
front50 for all pipelines.

Note that both before and after this change, when ManualEventHandler gets a pipeline from the
cache, it queries front50 for a single pipeline by id to make sure what it's triggering is
up to date.  As currently implemented, front50 queries its database by id to generate the
response.

Currently, deck triggers by pipeline application + name.  See for example
https://github.com/spinnaker/deck/blob/8abafe5c6a6680953aef7752b1a56cac093ec9ba/packages/core/src/pipeline/executions/executionGroup/ExecutionGroup.tsx#L138
and
https://github.com/spinnaker/deck/blob/8abafe5c6a6680953aef7752b1a56cac093ec9ba/packages/core/src/domain/IPipeline.ts#L86

With this PR, ManualEventHandler now queries a new endpoint in front50 by application and
name when there's no cached pipeline corresponding to a trigger.  Front50's database query
in this case is likely slightly more expensive that querying by id, but not enough to be
significant, especially since we're dealing with human-initiated events.

If the query by application and name fails, ManualEventTrigger then queries by pipeline
id, so triggers by pipeline id (if there are any) are less performant than before as they
involve another round trip to front50 to query by application and name first.

It does seem feasible, though not totally trivial, to change deck to trigger manual
execution by pipeline id.  If this happens, it would make sense to change the order of
queries in ManualEventHandler to query by id first.  That would restore the performance of
manual execution by deck what it was before this PR.